refactor(ProductActions): streamline product data fetching and caching
- Consolidated product data fetching logic by utilizing a caching mechanism.
- Updated getJsonFileData to improve error handling and data retrieval.
- Removed redundant file path references and optimized category fetching logic.
- Enhanced overall code organization for better maintainability.
