+++ lib/actions/product.actions.ts
@@ -4,44 +4,64 @@ import path from "path";
 import { getJsonFileData } from "@/lib/utils";
 // import { ARTE_PRODUCTS_FILE_PATH } from "../fetch-artegifts-data";
 
-export async function getAllProducts() {
-  // const DATA_FILE_PATH = path.join(process.cwd(), '_data/products-20-03-25.json');
-  const DATA_FILE_PATH = '../_data/products-27-03-25.json';
-  const { data } = await getJsonFileData(DATA_FILE_PATH) ?? {};
 
-  return data?.item ?? [];
+export async function getAllProducts() {
+  const jsonData = await getJsonFileData();
+  return jsonData?.data?.item ?? [];
 }
 
-/**
- * Retrieves a list of unique product categories from the available products.
- *
- * @return {string[]} An array of unique product categories
- */
 export async function getAllProductCategories() {
-  // Fetch all products
   const products = await getAllProducts();
 
-  // Define a function that takes a category string and returns the first part of it
-  // The category string is in the format "category|subcategory"
-  const getCategory = (category: string) => {
-    // Split the category string into an array of two strings
-    const [categoryName, subcategoryName] = category[0].split('|');
-    // Return the first string (the category name)
-    return categoryName;
-  };
+  const categories = new Set<string>();
 
-  // Create a Set to store the unique categories
-  const categories = new Set();
-
-  // Loop through each product and add its category to the Set
-  products?.map((product: any) => {
-    const categoryName = getCategory(product.category);
+  for (const product of products) {
+    if (!product?.category) continue;
+    const [categoryName] = product.category[0].split('|');
     categories.add(categoryName);
-  });
+  }
 
-  // Convert the Set to an array and return it
   return Array.from(categories);
 }
+
+// export async function getAllProducts() {
+//   // const DATA_FILE_PATH = path.join(process.cwd(), '_data/products-20-03-25.json');
+//   const DATA_FILE_PATH = '../_data/products-27-03-25.json';
+//   const { data } = await getJsonFileData(DATA_FILE_PATH) ?? {};
+
+//   return data?.item ?? [];
+// }
+
+// /**
+//  * Retrieves a list of unique product categories from the available products.
+//  *
+//  * @return {string[]} An array of unique product categories
+//  */
+// export async function getAllProductCategories() {
+//   // Fetch all products
+//   const products = await getAllProducts();
+
+//   // Define a function that takes a category string and returns the first part of it
+//   // The category string is in the format "category|subcategory"
+//   const getCategory = (category: string) => {
+//     // Split the category string into an array of two strings
+//     const [categoryName, subcategoryName] = category[0].split('|');
+//     // Return the first string (the category name)
+//     return categoryName;
+//   };
+
+//   // Create a Set to store the unique categories
+//   const categories = new Set();
+
+//   // Loop through each product and add its category to the Set
+//   products?.map((product: any) => {
+//     const categoryName = getCategory(product.category);
+//     categories.add(categoryName);
+//   });
+
+//   // Convert the Set to an array and return it
+//   return Array.from(categories);
+// }
 
 export async function getProductsByLimit(limit: number) {
   // Use a more efficient way to get a subset of the products array
